The crunch: a rounded tick and an interpolator reading off the end
Ty: "the half time one does not seem to be functioning correctly it sounds
the same but its being cut off ... lots of crunchyness in the sounds too ...
it seems the diagonal lines on time changing is causing this buzz crunchy
sound." Two faults, both measured before anything was touched, and neither of
them in the DSP.

**The song's position was rounded before the effect saw it.** A tick is an
i64 on `TransportSnapshot`, and at 120 bpm a tick is twenty-five samples
while the snapshot is rebuilt once a block — so the phase snapped backwards
by up to a whole tick at every boundary. On a flat curve that is invisible.
On a slope it is a jump in the read position: twelve samples at half speed,
three hundred and seventy-five times a second, which is exactly the diagonals
he pointed at. A clean 500 Hz tone's biggest kink is 0.0043; through the old
chain it was 1.31. It carries `f64` from `tick_at_exact` now, and `tick_at`
stays rounded for the grids and markers that want a whole tick.

**The interpolator read off the end of the memory.** A four-point Hermite
wants two samples either side and the ring has none past the write head, and
`read_ring` filled those taps with zeros. A cubic through that cliff
overshoots: one sample at 0.664 where every neighbour is 0.625, every time a
curve's read touches live at a fractional position — which is what every
segment faster than 1x does on its way back to the present. Five rows did it
twice a bar and it was every over-unity peak in the bank. The taps are held
at the edge, which is what a resampler does at the end of a buffer.

Three presets also did nothing whatever. Push, Rushed and one beat of Drunk
were drawn reading the future; the clamp eats that sample by sample, and a
preset that does nothing has no symptom because a wire sounds fine. The two
that mean it buy the beat of look-ahead and say so in their notes.

And Half-time was doing what it was drawn to do — beats 3 and 4 replaying
beats 1 and 2 — which is why it sounded the same. The name promised a speed.
It is Half-time Feel, beside Tape's Half Speed.

Five tests, two of them sweeps over the whole bank: nothing may move the read
head without a crossfade, nothing may pass unity, and every drawn scene must
differ audibly from an empty one. Both sweeps failed the first time they ran.
Aliasing on speed-up was measured and left: at every rate in the bank the
folded images land above 20 kHz.

+++ crates/fontelle-engine/src/nodes.rs
@@ -1566,7 +1566,7 @@ impl AudioNode for EffectNode {
         ) = (&mut self.state, &self.config)
         {
             let music = fontelle_types::MusicalTime {
-                tick: ctx.transport.position_tick as f64,
+                tick: ctx.transport.position_tick,
                 ticks_per_sample: ctx.transport.ticks_per_sample,
                 beats_per_bar: ctx.transport.beats_per_bar.max(1),
                 bpm: ctx.transport.bpm,
